add warning and checks at compilation, modify sources according to new rules

+++ source/adc.hpp
@@ -3,7 +3,7 @@
 #include "hardwareConf.hpp"
 
 
-constexpr float rad2deg(const float rad) {return rad * 180 / M_PI;}
+constexpr float rad2deg(const float rad) {return rad * 180.0f / static_cast<float>(M_PI);}
 
 class Adc : public WorkerThread<TH_ADC::threadStackSize, Adc> {
 public:
